refactor: shared component
