+++ packages/widgets-internal/farm/components/FarmTable/FarmTokenInfo.tsx
@@ -1,14 +1,9 @@
 import { useTranslation } from "@pancakeswap/localization";
-import { Box, Flex, Link, Row, Skeleton, Text, TooltipText, WarningIcon, useTooltip } from "@pancakeswap/uikit";
+import { Flex, Row, Skeleton, Text } from "@pancakeswap/uikit";
 import { useMemo } from "react";
 import { styled } from "styled-components";
-import { isMobile } from "react-device-detect";
 import { FarmTableFarmTokenInfoProps } from "../../types";
-
-const InlineLink = styled(Link)`
-  display: inline;
-  margin-left: 4px;
-`;
+import MerklNotice from "../MerklNotice";
 
 const Container = styled.div`
   padding-left: 16px;
@@ -29,31 +24,6 @@ const TokenWrapper = styled.div`
   }
 `;
 
-const MerklTooltip: React.FC = () => {
-  const { t } = useTranslation();
-  return (
-    <>
-      <Box>
-        <Text display="inline">
-          <p>
-            {t("Incentives have moved to")}
-            <InlineLink
-              external
-              display="inline"
-              href="https://merkl.angle.money/?times=active%2Cfuture%2C&phrase=RETH&chains=1%2C"
-            >
-              {t("Merkl")}
-            </InlineLink>
-            , {t("and are now claimable without staking your LP token.")}
-          </p>
-          <br />
-          {t("Continue seeding your liquidity on PancakeSwap to accrue the rewards!")}
-        </Text>
-      </Box>
-    </>
-  );
-};
-
 const Farm: React.FunctionComponent<React.PropsWithChildren<FarmTableFarmTokenInfoProps>> = ({
   label,
   isReady,
@@ -61,11 +31,6 @@ const Farm: React.FunctionComponent<React.PropsWithChildren<FarmTableFarmTokenIn
   children,
 }) => {
   const { t } = useTranslation();
-  const { tooltip, tooltipVisible, targetRef } = useTooltip(<MerklTooltip />, {
-    placement: "top-start",
-    tooltipOffset: [-20, 10],
-    trigger: isMobile ? "focus" : "hover",
-  });
 
   const handleRenderFarming = useMemo(() => {
     if (isStaking) {
@@ -97,13 +62,8 @@ const Farm: React.FunctionComponent<React.PropsWithChildren<FarmTableFarmTokenIn
         {handleRenderFarming}
         <Row gap="sm">
           <Text bold>{label}</Text>
-          <TooltipText ref={targetRef}>
-            <Text lineHeight={0}>
-              <WarningIcon color="warning" />
-            </Text>
-          </TooltipText>
+          <MerklNotice />
         </Row>
-        {tooltipVisible ? tooltip : null}
       </div>
     </Container>
   );
